TextView: Fix text cursor to handle formatting correctly
Previously, the text cursor was *repainted* each blink (background, then text),
which caused numerous issues (!) when grapheme under the cursor is formatted.

Now, we paint the text cursor onto a QPixmap twice -- visible and invisible,
then use these QPixmaps when repainting the text cursor.

To be clear, this has introduced a *lot* of new code!

Other:
* TextViewGraphemePosition & GraphemeFinder
** Add: bool isEndOfLine
   Not strictly required, but nice to know during debugging.  :)